requirements.txt
cffi==1.10.0
pycparser==2.17
PyNaCl==1.1.2
six==1.10.0
